added all tests to test file
