--- types/mercadopago-sdk-js/test/bricks.ts
@@ -1,18 +1,0 @@
-import { mpInstance } from "../mercadopago-sdk-js-tests";
-
-const brickBuilder = mpInstance.bricks({})
-
-brickBuilder.isInitialized();
-
-brickBuilder.create('cardPayment', 'container', {
-    initialization: {
-        amount: 100
-    },
-    callbacks: {
-        onSubmit: (formData, additionalData) => {
-            return new Promise(() => {
-                console.log(formData, additionalData)
-            })
-        }
-    }
-})
